app/requirements.txt
alembic==1.7.7
boto3==1.21.44
botocore==1.24.44
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cycler==0.11.0
Flask==2.1.1
Flask-Cors==3.0.10
Flask-Login==0.6.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Toastr==0.5.7
fonttools==4.32.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
kiwisolver==1.4.2
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
psycopg2==2.9.3
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.35
urllib3==1.26.9
Werkzeug==2.1.1
